Move function is_mgmt_vrf_enabled
