+++ tests/route/test_forced_mgmt_route.py
@@ -8,7 +8,7 @@ from tests.common.helpers.assertions import pytest_assert
 from tests.common.utilities import wait_until
 from tests.common.utilities import backup_config, restore_config, \
         reload_minigraph_with_golden_config
-from tests.syslog.syslog_utils import is_mgmt_vrf_enabled
+from tests.common.helpers.syslog_helpers import is_mgmt_vrf_enabled
 
 pytestmark = [
     pytest.mark.disable_loganalyzer,
